# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a API documentation for use marketplace database.",
        "title": "Swagger Marketplace API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:6060",
    "basePath": "/api/v1",
    "paths": {
        "/carts/": {
            "post": {
                "description": "post cart data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carts"
                ],
                "summary": "Insert cart data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your admin JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.CartResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    }
                }
            }
        },
        "/carts/:id": {
            "get": {
                "description": "get cart data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carts"
                ],
                "summary": "Get cart data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your admin JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.CartResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    }
                }
            },
            "put": {
                "description": "put cart data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carts"
                ],
                "summary": "Edit cart data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your admin JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.CartResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete cart data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carts"
                ],
                "summary": "Discard cart data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your admin JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    }
                }
            }
        },
...
        "/users/:id/carts": {
            "get": {
                "description": "get all carts from userID data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carts"
                ],
                "summary": "Get user carts data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Insert your admin JWT token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.CartsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_gabrielssssssssss_marketplace-telegram_internal_model.Cart": {
            "type": "object",
            "properties": {
                "cart_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_gabrielssssssssss_marketplace-telegram_internal_model.CartResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Cart"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_gabrielssssssssss_marketplace-telegram_internal_model.CartsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_gabrielssssssssss_marketplace-telegram_internal_model.Cart"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_gabrielssssssssss_marketplace-telegram_internal_model.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}